made cache utility generic
